fix(rust-api): restore web_assets module to crate root (sc-1636)
The generation.rs extraction (#224) accidentally swept the
`#[cfg(feature = "embed-web")] mod web_assets` block into generation.rs
because it sat between create_video_job and app_fallback in the original
layout. That nested it as `crate::generation::web_assets`, so the
embed-web build broke: app_fallback (at the root) and tests.rs both
reference it via the crate-root path (`web_assets::serve`,
`super::web_assets::CONTENT_SECURITY_POLICY`).

The default/server/test builds never compile the embed-web path, so
`npm run rust:check` stayed green and only the Windows desktop installer
build (`api:build:embedded`, which compiles --features embed-web) failed.

Move the module back to the crate root next to app_fallback. Verified
with `cargo build -p sceneworks-rust-api --features embed-web`.
