+++ apps/rust-api/src/lib.rs
@@ -1452,6 +1452,70 @@ fn parse_single_byte_range(value: &str, total: u64) -> Option<(u64, u64)> {
     Some((start, end))
 }
 
+/// Embedded production web bundle (apps/web/dist), compiled in only under the
+/// `embed-web` feature so default/server/test builds need no web build.
+#[cfg(feature = "embed-web")]
+mod web_assets {
+    use axum::http::{header, StatusCode, Uri};
+    use axum::response::{IntoResponse, Response};
+    use rust_embed::RustEmbed;
+
+    #[derive(RustEmbed)]
+    #[folder = "../web/dist"]
+    struct WebAssets;
+
+    // The desktop shell navigates its privileged webview to this server, so the embedded
+    // UI runs from this origin and its CSP must come from here (tauri.conf.json only
+    // governs the bundled setup screen). Kept narrow: scripts only from this origin (the
+    // theme bootstrap was moved to /theme-init.js so no inline script is needed), Google
+    // Fonts allowed, images/media as self/data/blob, IPC for the Tauri webview. Same-origin
+    // API + SSE are covered by connect-src 'self'.
+    pub(super) const CONTENT_SECURITY_POLICY: &str = "default-src 'self'; \
+script-src 'self'; \
+style-src 'self' 'unsafe-inline' https://fonts.googleapis.com; \
+font-src 'self' https://fonts.gstatic.com data:; \
+img-src 'self' data: blob:; \
+media-src 'self' data: blob:; \
+connect-src 'self' ipc: http://ipc.localhost; \
+object-src 'none'; \
+base-uri 'self'; \
+frame-ancestors 'none'; \
+form-action 'self'";
+
+    pub(super) async fn serve(uri: Uri) -> Response {
+        let requested = uri.path().trim_start_matches('/');
+        let requested = if requested.is_empty() {
+            "index.html"
+        } else {
+            requested
+        };
+        if let Some(file) = WebAssets::get(requested) {
+            let mime = mime_guess::from_path(requested).first_or_octet_stream();
+            return (
+                [
+                    (header::CONTENT_TYPE, mime.as_ref()),
+                    (header::CONTENT_SECURITY_POLICY, CONTENT_SECURITY_POLICY),
+                ],
+                file.data.into_owned(),
+            )
+                .into_response();
+        }
+        // Single-page app: unknown non-API paths resolve to index.html so
+        // client-side deep links (e.g. project routes) load correctly.
+        match WebAssets::get("index.html") {
+            Some(index) => (
+                [
+                    (header::CONTENT_TYPE, "text/html; charset=utf-8"),
+                    (header::CONTENT_SECURITY_POLICY, CONTENT_SECURITY_POLICY),
+                ],
+                index.data.into_owned(),
+            )
+                .into_response(),
+            None => StatusCode::NOT_FOUND.into_response(),
+        }
+    }
+}
+
 /// Router fallback. With `embed-web`, non-API paths are served from the embedded
 /// web bundle (SPA fallback); API paths and all default-feature builds keep the
 /// existing JSON not-found behavior.
